move dump checks in a dedicated directory
