+++ test/IO/dump/definitions.hpp
@@ -1,0 +1,5 @@
+#define     COMPONENTS      3
+#define     DIMENSIONS      3
+//#define DEBUG
+
+#define     GEOMETRY        CARTESIAN
